Fix for PHP warnings at http://code.google.com/p/sloodle/issues/detail?id=110

+++ trunk/sloodle/lib/layout_profile.php
@@ -18,7 +18,7 @@
         var $courseid;
 
         var $entries = array();
-        var $originalentryids = array();
+        var $originalentries = array();
 
         function SloodleLayout($r = null) {
 
@@ -139,12 +139,14 @@
                $neededids[] = $entry->id;
             }
 
-            foreach($this->originalentries as $entry) {
-               $entryid = $entry->id;
-               if (!in_array($entryid,$neededids)) {
-                  $entry->delete();
+	    if (count($this->originalentries) > 0) {
+               foreach($this->originalentries as $entry) {
+                  $entryid = $entry->id;
+                  if (!in_array($entryid,$neededids)) {
+                     $entry->delete();
+                  }
                }
-            }
+	    }
 
            return true;
